컨트롤러 리팩토링: 서로 다른 컨트롤러에 산재해 있는 메서드를 취합합니다.
기존의 컨트롤러 메서드는 협업으로 인해 같은 도메인에 속한 컨트롤러라도,
서로 다른 컨트롤러 클래스에 속해 있는 경우가 있었습니다.
이를 도메인에 적절하게 정리합니다.
