+++ src/main/java/com/scit/iLog/controller/DashboardController.java
@@ -5,8 +5,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class DashboardController {
-    
-
     @GetMapping("/parentDashboard")
     public String parentDashboard() {
